v1.1-SNAPSHOT

+++ src/main/java/com/aaronr92/auroraproject/model/Particle.java
@@ -85,7 +85,7 @@ public class Particle {
         return offsetZ;
     }
 
-    public int getSize() {
+    public double getSize() {
         return size;
     }
 }
